+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Inputs inside a form group no longer have a background color by default; use `is
   - Added attr `is_monospace`
 - `checkbox` and `radio_button`
   - Changed internal structure
+- `radio_group`
+  - Added input styling from Primer ViewComponents radio button (keeping the initial size)
 - `select`:
   - Added wrapper element
   - Added attr `is_monospace`